+++ R/rstudent.R
@@ -28,7 +28,7 @@ rstudent.deconv <- function(model, ...) {
   r <- model$subclass$residuals
   weights <- model$subclass$weights
   if (!is.null(weights)) r <- r * weights
-  hat <- model$subclass$hat
+  hat <- hat(model)
   rdf <- nrow(r) - ncol(model$subclass$compensation)
   rss <- colSums(r^2)
   mse <- rss / nrow(r)
@@ -43,7 +43,7 @@ rstandard.deconv <- function(model, ...) {
   r <- model$subclass$residuals
   weights <- model$subclass$weights
   if (!is.null(weights)) r <- r * weights
-  hat <- model$subclass$hat
+  hat <- hat(model)
   rdf <- nrow(r) - ncol(model$subclass$compensation)
   rss <- colSums(r^2)
   mse <- rss / nrow(r)
@@ -57,7 +57,7 @@ cooks.distance.deconv <- function(model, ...) {
   r <- model$subclass$residuals
   weights <- model$subclass$weights
   if (!is.null(weights)) r <- r * weights
-  hat <- model$subclass$hat
+  hat <- hat(model)
   p <- ncol(model$subclass$compensation)
   rdf <- nrow(r) - p
   rss <- colSums(r^2)
@@ -71,7 +71,7 @@ rstudent_fit <- function(fit) {
   r <- fit$residuals
   weights <- fit$weights
   if (!is.null(weights)) r <- r * weights
-  hat <- fit$hat
+  hat <- hat_fit(fit)
   rdf <- nrow(r) - ncol(fit$compensation)
   rss <- colSums(r^2)
   mse <- rss / nrow(r)
@@ -84,12 +84,27 @@ cooks_distance_fit <- function(fit) {
   r <- fit$residuals
   weights <- fit$weights
   if (!is.null(weights)) r <- r * weights
-  hat <- fit$hat
+  hat <- hat_fit(fit)
   p <- ncol(fit$compensation)
   rdf <- nrow(r) - p
   rss <- colSums(r^2)
   mse <- rss / rdf
   (r / (1 - hat))^2 * outer(hat, mse, "/") / p
+}
+
+
+hat <- function(model) {
+  X <- model$subclass$X
+  Lv <- colSums(X^2)
+  iXTX <- model$subclass$compensation / Lv
+  diag(X %*% iXTX %*% t(X))
+}
+
+hat_fit <- function(fit) {
+  X <- fit$X
+  Lv <- colSums(X^2)
+  iXTX <- fit$compensation / Lv
+  diag(X %*% iXTX %*% t(X))
 }
 
 
@@ -102,23 +117,17 @@ cooks_distance_fit <- function(fit) {
 #' @param object a 'deconv' class object
 #' @param ... retained for class compatibility
 #' @param test bulk gene expression matrix assumed to be in raw counts
-#' @param arith_mean logical, whether to use arithmetic mean as gene signature
-#' @param use_filter logical, whether to use denoised signature matrix
 #' @returns Matrix of residuals.
 #' @export
-residuals.deconv <- function(object,
-                             test = NULL,
-                             arith_mean = FALSE, use_filter = FALSE, ...) {
+residuals.deconv <- function(object, ...,
+                             test = NULL) {
   if (is.null(test)) return(object$subclass$residuals)
   # recalculate residuals
-  if (is.null(arith_mean)) arith_mean <- object$call$arith_mean
-  if (is.null(use_filter)) use_filter <- object$call$use_filter
-  if (arith_mean) {
-    cellmat <- if (use_filter) object$mk$genemeans_ar_filter else object$mk$genemeans_ar
-    if (is.null(cellmat)) stop("arithmetic mean not available")
-  } else {
-    cellmat <- if (use_filter) object$mk$genemeans_filter else object$mk$genemeans
-  }
+  arith_mean <- eval(object$call$arith_mean) %||% FALSE
+  use_filter <- eval(object$call$use_filter) %||% TRUE
+  count_space <- eval(object$call$count_space) %||% TRUE
+  cellmat <- get_cellmat(object$mk, arith_mean, use_filter, sub = TRUE)
+  
   nm <- object$call$test
   .call <- match.call()
   if (nm != .call$test) {
@@ -128,6 +137,77 @@ residuals.deconv <- function(object,
   geneset <- intersect(rownames(cellmat), rownames(test))
   cellmat <- cellmat[geneset, ]
   test <- test[geneset, ]
-  cellmat <- 2^cellmat -1
+  if (count_space) cellmat <- 2^cellmat -1
   residuals_deconv(test, cellmat, object$subclass$output)
+}
+
+
+#' Standard errors of deconvoluted cell counts
+#' 
+#' Extracts standard errors of deconvoluted cell counts based on the row
+#' variance of weighted residuals.
+#' 
+#' @param model a 'deconv' class object
+#' @param type specifies standard error method. Default is "var.e", which we
+#'   recommend as being the most accurate. The alternative options include OLS
+#'   (ordinary least squares) and several heteroscedasticity-consistent methods,
+#'   HC0, HC2 and HC3.
+#' @returns with `type = "var.e"`, a vector of standard errors of cell counts
+#'   for each cell subclass. For all other options, a matrix of SE for every
+#'   bulk sample and every cell subclass.
+#' @seealso [deconvolute()]
+#' @export
+se <- function(model, type = c("var.e", "OLS", "OLS2", "HC0", "HC2", "HC3")) {
+  if (!inherits(model, "deconv")) stop("not a 'deconv' class object")
+  type <- match.arg(type)
+  X <- model$subclass$X
+  var.e <- model$subclass$var.e
+  Lv <- colSums(X^2)
+  iXTX <- model$subclass$compensation / Lv
+  resvar <- model$subclass$resvar
+  XTXse <- crossprod(X, var.e * X)
+  
+  if (type == "var.e") {
+    # var(beta) = (X' X)^-1 (X diag(e^2) X') (X' X)^-1
+    se <- sqrt(diag(iXTX %*% XTXse %*% t(iXTX)))
+    return(se)
+  } else if (type == "OLS") {
+    diag_XTX <- diag(model$subclass$compensation) / Lv
+    se <- sqrt(resvar %*% t(diag_XTX))
+  } else if (type == "OLS2") {
+    # based on van Wieringen
+    XTX <- crossprod(X)
+    diag2 <- diag(iXTX %*% XTX %*% t(iXTX))
+    se <- sqrt(resvar %*% t(diag2))
+  } else {
+    # heteroscedasticity-consistent SE, HC0-3
+    r <- model$subclass$residuals
+    weights <- model$subclass$weights
+    if (!is.null(weights)) r <- r * weights
+    hat <- hat(model)
+    omega <- switch(type, "HC0" = function(i) i^2,
+                    "HC2" = function(i) i^2 / (1 - hat),
+                    "HC3" = function(i) i^2 / (1 - hat)^2)
+    
+    se <- t(apply(r, 2, function(i) {
+      XTXse <- crossprod(X, omega(i) * X)
+      sqrt(diag(iXTX %*% XTXse %*% t(iXTX)))
+    }))
+  }
+  rownames(se) <- rownames(model$subclass$output)
+  se
+}
+
+
+#' Compute condition number of deconvolution model
+#' 
+#' Computes the condition number of the spillover matrix from a deconvolution
+#' model.
+#' 
+#' @param z a 'deconv' class object
+#' @param ... arguments passed to [kappa()]
+#' @returns The condition number.
+#' @export
+kappa.deconv <- function(z, ...) {
+  kappa(z$subclass$spillover, ...)
 }